# contracts/openapi/clipshare-v1.openapi.json
{
  "components": {
    "schemas": {
      "Body_uploadFile": {
        "properties": {
          "encrypted": {
            "default": false,
            "title": "Encrypted",
            "type": "boolean"
          },
          "expiry": {
            "default": "24h",
            "enum": [
              "1h",
              "24h",
              "7d",
              "forever"
            ],
            "title": "Expiry",
            "type": "string"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "max_views": {
            "default": "",
            "title": "Max Views",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_uploadFile",
        "type": "object"
      },
      "FileCreatedResponse": {
        "description": "文件上传成功响应（201）。",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{6}Z$",
            "title": "Created At",
            "type": "string"
          },
          "encrypted": {
            "title": "Encrypted",
            "type": "boolean"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{6}Z$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "max_views": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Views"
          },
          "original_name": {
            "title": "Original Name",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "code",
          "url",
          "original_name",
          "size_bytes",
          "encrypted",
          "expires_at",
          "max_views",
          "created_at"
        ],
        "title": "FileCreatedResponse",
        "type": "object"
      },
      "FileReadResponse": {
        "description": "文件元数据响应（200）：前端据此渲染文件卡片与下载/预览按钮。",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{6}Z$",
            "title": "Created At",
            "type": "string"
          },
          "encrypted": {
            "title": "Encrypted",
            "type": "boolean"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{6}Z$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "kind": {
            "const": "file",
            "default": "file",
            "title": "Kind",
            "type": "string"
          },
          "original_name": {
            "title": "Original Name",
            "type": "string"
          },
          "preview_available": {
            "title": "Preview Available",
            "type": "boolean"
          },
          "remaining_views": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining Views"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "code",
          "original_name",
          "size_bytes",
          "encrypted",
          "content_type",
          "preview_available",
          "expires_at",
          "remaining_views",
          "created_at"
        ],
        "title": "FileReadResponse",
        "type": "object"
      },
      "ProblemDetail": {
        "additionalProperties": false,
        "description": "客户端冻结的 RFC 9457 风格错误契约。",
        "properties": {
          "detail": {
            "description": "本次错误的具体说明",
            "title": "Detail",
            "type": "string"
          },
          "status": {
            "description": "HTTP 状态码",
            "maximum": 599,
            "minimum": 400,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "面向用户的简短标题",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "稳定机器错误码",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "轻量级云剪切板分享系统 API",
    "title": "ClipShare",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/files": {
      "post": {
        "description": "流式上传文件：净化文件名 → 白名单 415 → 参数校验 422 → 流式落盘。\n\n落盘超限 413 由存储层中断并自清理半成品；服务层加密上限兜底 422 时\n路由在进入数据库提交前失败时补偿删除已落盘文件。提交一旦开始，其结果可能\n因连接中断而无法确认；此时优先保留文件供后续引用核验/回收，绝不能误删一个\n事实上已经提交的数据库记录所引用的内容。",
        "operationId": "uploadFile",
        "parameters": [
          {
            "description": "可选的创建请求幂等键；8–128 位 ASCII，只允许字母、数字和 ._~:/+-。服务端仅保存 SHA-256 摘要；格式错误返回 400 idempotency_key_invalid。",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "可选的创建请求幂等键；8–128 位 ASCII，只允许字母、数字和 ._~:/+-。服务端仅保存 SHA-256 摘要；格式错误返回 400 idempotency_key_invalid。",
              "maxLength": 128,
              "minLength": 8,
              "pattern": "^[A-Za-z0-9._~:/+\\-]{8,128}$",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadFile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileCreatedResponse"
                }
              }
            },
            "description": "上传成功或幂等重放",
            "headers": {
              "Cache-Control": {
                "description": "固定为 no-store，禁止缓存绕过过期和访问次数",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "Idempotency-Replayed": {
                "description": "携带幂等键时，true 表示返回既有结果",
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "请求或幂等键格式错误",
            "headers": {
              "Cache-Control": {
                "description": "固定为 no-store，禁止缓存绕过过期和访问次数",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "幂等键冲突",
            "headers": {
              "Cache-Control": {
                "description": "固定为 no-store，禁止缓存绕过过期和访问次数",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            }
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "请求内容过大",
            "headers": {
              "Cache-Control": {
                "description": "固定为 no-store，禁止缓存绕过过期和访问次数",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            }
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "媒体类型或文件类型不受支持",
            "headers": {
              "Cache-Control": {
                "description": "固定为 no-store，禁止缓存绕过过期和访问次数",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "请求参数校验失败",
            "headers": {
              "Cache-Control": {
                "description": "固定为 no-store，禁止缓存绕过过期和访问次数",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "请求速率超限",
            "headers": {
              "Cache-Control": {
                "description": "固定为 no-store，禁止缓存绕过过期和访问次数",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "客户端再次尝试前应等待的十进制秒数",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "服务器内部错误",
            "headers": {
              "Cache-Control": {
                "description": "固定为 no-store，禁止缓存绕过过期和访问次数",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "上传文件",
        "tags": [
          "files"
        ],
        "x-clipshare-auto-retry": "idempotency-key",
        "x-clipshare-consumes-view": false,
        "x-clipshare-max-encrypted-marker-chars": 16777216,
        "x-clipshare-max-encrypted-plaintext-bytes": 10485760,
        "x-clipshare-max-upload-bytes": 104857600
      }
    },
    "/api/v1/files/{code}": {
      "get": {
        "description": "文件元数据：不消耗预览/下载次数（与文本分享二维码同语义）。\n\n不存在 404；已过期 410（服务层顺带懒删磁盘文件）。",
        "operationId": "getFileMetadata",
        "parameters": [
          {
            "description": "1–8 位 Base62 分享短码",
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "description": "1–8 位 Base62 分享短码",
              "maxLength": 8,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9]+$",
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileReadResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "description": "固定为 no-store，禁止缓存绕过过期和访问次数",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "资源不存在",
            "headers": {
              "Cache-Control": {
                "description": "固定为 no-store，禁止缓存绕过过期和访问次数",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            }
          },
          "410": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "资源已过期、耗尽或不可再获取",
            "headers": {
              "Cache-Control": {
                "description": "固定为 no-store，禁止缓存绕过过期和访问次数",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "请求参数校验失败",
            "headers": {
              "Cache-Control": {
                "description": "固定为 no-store，禁止缓存绕过过期和访问次数",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "请求速率超限",
            "headers": {
              "Cache-Control": {
                "description": "固定为 no-store，禁止缓存绕过过期和访问次数",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "客户端再次尝试前应等待的十进制秒数",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "服务器内部错误",
            "headers": {
              "Cache-Control": {
                "description": "固定为 no-store，禁止缓存绕过过期和访问次数",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "读取文件元数据",
        "tags": [
          "files"
        ],
        "x-clipshare-auto-retry": "safe",
        "x-clipshare-consumes-view": false
      }
    },
    "/api/v1/files/{code}/download": {
      "get": {
        "description": "下载文件：消耗一次下载计数，FileResponse 流式输出。\n\n中文文件名经 RFC 5987 filename* 编码（Starlette 内建处理）；\n磁盘文件缺失（记录尚在）→ 410 file_content_missing。",
        "operationId": "downloadFile",
        "parameters": [
          {
            "description": "1–8 位 Base62 分享短码",
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "description": "1–8 位 Base62 分享短码",
              "maxLength": 8,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9]+$",
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "文件二进制流",
            "headers": {
              "Cache-Control": {
                "description": "固定为 no-store，禁止缓存绕过过期和访问次数",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "Content-Disposition": {
                "description": "RFC 5987 编码的原始下载文件名",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "资源不存在",
            "headers": {
              "Cache-Control": {
                "description": "固定为 no-store，禁止缓存绕过过期和访问次数",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            }
          },
          "410": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "资源已过期、耗尽或不可再获取",
            "headers": {
              "Cache-Control": {
                "description": "固定为 no-store，禁止缓存绕过过期和访问次数",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "请求参数校验失败",
            "headers": {
              "Cache-Control": {
                "description": "固定为 no-store，禁止缓存绕过过期和访问次数",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "请求速率超限",
            "headers": {
              "Cache-Control": {
                "description": "固定为 no-store，禁止缓存绕过过期和访问次数",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "客户端再次尝试前应等待的十进制秒数",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "服务器内部错误",
            "headers": {
              "Cache-Control": {
                "description": "固定为 no-store，禁止缓存绕过过期和访问次数",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "下载文件",
        "tags": [
          "files"
        ],
        "x-clipshare-auto-retry": "forbidden",
        "x-clipshare-consumes-view": true
      }
    },
...
  }
}